requirements.txt
environs==9.5.0
python-telegram-bot==13.15
Jinja2==2.10.1
openpyxl==3.0.10
phonenumbers==8.13.4
django==4.2
django-phonenumber-field==7.0.2
Pillow==9.5.0
